+++ duplocloud/resource_duplo_gcp_sql_database.go
@@ -76,7 +76,30 @@ func gcpSqlDBInstanceSchema() map[string]*schema.Schema {
 			Optional:    true,
 			Default:     true,
 		},
+		"root_password": {
+			Description: "root password for specific database versions",
+			Type:        schema.TypeString,
+			Optional:    true,
+		},
 	}
+}
+
+func checkPasswordNeeded(d *schema.ResourceData) bool {
+
+	// Check the value of dependent_field
+
+	dependentFieldValue := d.Get("database_version").(string)
+	mp := map[string]bool{
+		"SQLSERVER_2017_STANDARD":   true,
+		"SQLSERVER_2017_ENTERPRISE": true,
+		"SQLSERVER_2017_EXPRESS":    true,
+		"SQLSERVER_2017_WEB":        true,
+		"SQLSERVER_2019_STANDARD":   true,
+		"SQLSERVER_2019_ENTERPRISE": true,
+		"SQLSERVER_2019_EXPRESS":    true,
+		"SQLSERVER_2019_WEB":        true,
+	}
+	return mp[dependentFieldValue]
 }
 func resourceGcpSqlDBInstance() *schema.Resource {
 	return &schema.Resource{
@@ -137,10 +160,12 @@ func resourceGcpSqlDBInstanceRead(ctx context.Context, d *schema.ResourceData, m
 // CREATE resource
 func resourceGcpSqlDBInstanceCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	log.Printf("[TRACE] resourceGcpSqlDBInstanceCreate ******** start")
-
 	// Create the request object.
 	rq := expandGcpSqlDBInstance(d)
 	tenantID := d.Get("tenant_id").(string)
+	if checkPasswordNeeded(d) && rq.RootPassword == "" {
+		return diag.Errorf("root password is mandatory for database version %s ", d.Get("database_version").(string))
+	}
 
 	c := m.(*duplosdk.Client)
 	fullName, clientErr := c.GetDuploServicesName(tenantID, rq.Name)
@@ -299,6 +324,7 @@ func expandGcpSqlDBInstance(d *schema.ResourceData) *duplosdk.DuploGCPSqlDBInsta
 		Tier:            d.Get("tier").(string),
 		DataDiskSizeGb:  d.Get("disk_size").(int),
 		ResourceType:    duplosdk.DuploGCPDatabaseInstanceResourceType,
+		RootPassword:    d.Get("root_password").(string),
 	}
 	if v, ok := d.GetOk("labels"); ok && !isInterfaceNil(v) {
 		rq.Labels = map[string]string{}